requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
ca-certificates=2019.11.28=hecc5488_0
certifi=2019.11.28=py38_0
cffi=1.13.2=py38h8022711_0
cryptography=2.8=py38h72c5cf5_1
future=0.18.2=py38_0
idna=2.8=py38_1000
ld_impl_linux-64=2.33.1=h53a641e_7
libffi=3.2.1=he1b5a44_1006
libgcc-ng=9.2.0=hdf63c60_0
libstdcxx-ng=9.2.0=hdf63c60_0
ncurses=6.1=hf484d3e_1002
openssl=1.1.1d=h516909a_0
pip=19.3.1=py38_0
pycparser=2.19=py38_1
pysocks=1.7.1=py38_0
python=3.8.0=h357f687_5
python-telegram-bot=12.2.0=py_0
readline=8.0=hf8c457e_0
setuptools=42.0.2=py38_0
six=1.13.0=py38_0
sqlite=3.30.1=hcee41ef_0
tk=8.6.10=hed695b0_0
tornado=6.0.3=py38h516909a_0
ujson=1.35=py38h516909a_1001
wheel=0.33.6=py38_0
xz=5.2.4=h14c3975_1001
zlib=1.2.11=h516909a_1006
